alipay-miniprogram-server/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
PyJWT==2.8.0
PyMySQL==1.1.0
cryptography==42.0.0
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==3.0.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
gunicorn==21.2.0
